// src/app/shared/models/driver/driver-param.model.ts
export class DriverParamModel {
  driverId: string;
  driverFirstName: string;
  driverLastName: string;
  driverLongitude: string;
  driverLatitude: string;
  driverStatus: string;
  driverEmail: string;
  driverRating: string;
  driverJobStatus: string;
  driverIdLink: string;
  driverIdBackLink: string;
  driverProfileImageLink: string;
  packageStatus: string;
  createdBy: string;
  createdOn: string;
  modifiedBy: string;
  modifiedOn: string;
  driverContact: string;
  city: string;
  state: string;
  country: string;
  pinCode: string;
  driverColorCode: string;
}